More test cases and MPI support in progress

+++ art/random_generator.hh
@@ -20,7 +20,7 @@ namespace art_modern {
 
     class Rprob {
     public:
-        Rprob(float pe_frag_dist_mean, float pe_frag_dist_std_dev, int read_length);
+        Rprob(double pe_frag_dist_mean, double pe_frag_dist_std_dev, int read_length);
         double r_prob();
         int insertion_length();
         char rand_base();
@@ -33,8 +33,8 @@ namespace art_modern {
     private:
 #if defined(USE_STL_RANDOM)
         std::mt19937 gen_;
-        std::uniform_real_distribution<float> dis_;
-        std::normal_distribution<float> insertion_length_gaussian_;
+        std::uniform_real_distribution<double> dis_;
+        std::normal_distribution<double> insertion_length_gaussian_;
         std::uniform_int_distribution<int> base_;
         std::uniform_int_distribution<int> strand_;
         std::uniform_int_distribution<int> quality_less_than_10_;
@@ -43,8 +43,8 @@ namespace art_modern {
         std::uniform_int_distribution<int> pos_on_read_not_head_and_tail_;
 #elif defined(USE_BOOST_RANDOM)
         boost::mt19937 gen_;
-        boost::uniform_real<float> dis_;
-        boost::normal_distribution<float> insertion_length_gaussian_;
+        boost::uniform_real<double> dis_;
+        boost::normal_distribution<double> insertion_length_gaussian_;
         boost::uniform_smallint<int> base_;
         boost::uniform_smallint<int> strand_;
         boost::uniform_smallint<int> quality_less_than_10_;
@@ -56,8 +56,8 @@ namespace art_modern {
 //    sycl::device rd_;
 //    sycl::queue queue_;
 //    oneapi::mkl::rng::mt19937 gen_;
-//    oneapi::mkl::rng::uniform<float, oneapi::mkl::rng::uniform_method::by_default> dis_;
-//    oneapi::mkl::rng::gaussian<float, oneapi::mkl::rng::gaussian_method::by_default> insertion_length_gaussian_;
+//    oneapi::mkl::rng::uniform<double, oneapi::mkl::rng::uniform_method::by_default> dis_;
+//    oneapi::mkl::rng::gaussian<double, oneapi::mkl::rng::gaussian_method::by_default> insertion_length_gaussian_;
 //    oneapi::mkl::rng::uniform<int, oneapi::mkl::rng::uniform_method::by_default> base_;
 //    oneapi::mkl::rng::uniform<int, oneapi::mkl::rng::uniform_method::by_default> strand_;
 //    oneapi::mkl::rng::uniform<int, oneapi::mkl::rng::uniform_method::by_default> quality_less_than_10_;
@@ -65,8 +65,8 @@ namespace art_modern {
 #elif defined(USE_GSL_RANDOM)
         const gsl_rng_type* T;
         gsl_rng* r;
-        float pe_frag_dist_mean_;
-        float pe_frag_dist_std_dev_;
+        double pe_frag_dist_mean_;
+        double pe_frag_dist_std_dev_;
         int read_length_;
 #endif
     };
